Move the files submodule over to snippets
It only depends on pathlib and seems generically useful to me, so off to snippets it goes!
