+++ tests/unit/test_node.py
@@ -4,7 +4,7 @@ import os
 import unittest
 
 from pyiron_workflow.channels import InputData, OutputData, NotData
-from pyiron_workflow.files import DirectoryObject
+from pyiron_workflow.snippets.files import DirectoryObject
 from pyiron_workflow.interfaces import Executor
 from pyiron_workflow.io import Inputs, Outputs
 from pyiron_workflow.node import Node
